Move update functions to function dir
