+++ src/components/aio/aioReplacementValuesDisplay.tsx
@@ -1,4 +1,4 @@
-import { fromHtml, toHtml } from 'components/functions';
+import { fromHtml, toHtml } from '../functions';
 import React, { useCallback, useState } from 'react';
 import { AioExternalReplacements, AioReplacement, AioReplacementValues } from './aioInterface';
 import { AioReplacementList } from './aioReplacementList';
