+++ src/components/functions/index.ts
@@ -9,6 +9,8 @@ import { objEqual } from "./objEqual";
 import { repeatHeaders } from "./repeatHeaders";
 import { repeatRows } from "./repeatRows";
 import { fromHtml, toHtml } from "./tofromHtml";
+import { updateReplacementVersion, updateReplToExtl } from "./updateReplacementVersion";
+import { updateTableDataVersion } from "./updateTableDataVersion";
 
 export {
   // assignSubListLevel,
@@ -23,4 +25,7 @@ export {
   repeatHeaders,
   repeatRows,
   toHtml,
+  updateReplacementVersion,
+  updateReplToExtl,
+  updateTableDataVersion,
 };
