+++ src/components/index.ts
@@ -4,6 +4,7 @@ import { AifBlockLine, AifLineType, AsupInternalBlock } from "./aif";
 import { AioDropSelect, AioExpander, AioExternalReplacements, AioIconButton, AioOption, AioReplacement, AioReplacementDisplay, AioReplacementValues, AioReplacementValuesDisplay, AioString } from "./aio";
 import { AitCellData, AitRowData, AitRowGroupData, AitTableData, AsupInternalTable } from "./ait";
 import { AsupInternalWindow } from "./aiw";
+import { updateReplacementVersion, updateReplToExtl, updateTableDataVersion } from "./functions";
 
 export {
   AsupInternalEditor,
@@ -24,4 +25,10 @@ export type { AieStyleMap };
 export type { AifBlockLine };
 export type { AioExternalReplacements, AioOption, AioReplacement, AioReplacementValues };
 export type { AitCellData, AitRowData, AitRowGroupData, AitTableData };
+
+export {
+  updateReplacementVersion,
+  updateReplToExtl,
+  updateTableDataVersion,
+}
