+++ src/index.ts
@@ -1,5 +1,1 @@
-import { updateReplacementVersion, updateReplToExtl } from 'pages/updateReplacementVersion';
-import { updateTableDataVersion } from 'pages/updateTableDataVersion';
-
 export * from './components';
-export { updateReplacementVersion, updateTableDataVersion, updateReplToExtl };
