+++ src/pages/ListPage.tsx
@@ -1,8 +1,8 @@
-import { AioExternalReplacements } from 'components/aio/aioInterface';
-import { newExternalReplacements } from 'components/functions/newExternalReplacements';
+import { AioExternalReplacements } from '../components/aio/aioInterface';
+import { newExternalReplacements } from '../components/functions/newExternalReplacements';
 import React, { useCallback, useRef, useState } from 'react';
 import { AioReplacementDisplay, AioString } from '../components';
-import { updateReplToExtl } from './updateReplacementVersion';
+import { updateReplToExtl } from '../components/functions/updateReplacementVersion';
 
 export const ListPage = (): JSX.Element => {
